requirements.txt
aiogram>=3.4,<4
python-dotenv>=1.0
Pillow>=10.0

# Для реальной генерации (опционально, тяжёлые зависимости):
# torch>=2.1
# diffusers>=0.38,<0.40
# transformers>=4.49,<5     # diffusers 0.38+ требует Dinov2WithRegistersConfig
# accelerate>=1.0           # peft 0.15+ требует clear_device_cache
# safetensors>=0.4
# peft>=0.15
# sentencepiece>=0.2        # требуется FLUX
